Edit payroll
Made working hours and time off nullable.
They are not known when creating paycheck at the start of the month.

Removed defualt date is the TRUNC command does not work. We would have
to use trim(leading '0' from to_char(checked_date, 'mm-yyyy')), but I do
not think this is necessary and may be too complicated.